Split tests into seperate files
